Peak spikes at 0:00
A typo on the normalize subroutine causes peak spikes at 0:00 each day. The data read from the vnstat command that is in the KiB range its been gathered in PiB instead of KiB because of the typo. The typo is on line 69 where is read as Kib instead of KiB.